move block copy button location
