+++ apps/www/src/lib/components/docs/block-chunk.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
 	import type { HTMLAttributes } from "svelte/elements";
-	import BlockCopyCodeButton from "../block-copy-code-button.svelte";
+	import BlockCopyCodeButton from "./block-copy-code-button.svelte";
 	import { cn, getLiftMode } from "$lib/utils.js";
 	import type { RawBlockChunk } from "$lib/blocks.js";
